+++ lib/commercetools-api/src/Models/Cart/CartSetCustomLineItemTaxAmountActionBuilder.php
@@ -39,6 +39,12 @@ final class CartSetCustomLineItemTaxAmountActionBuilder implements Builder
     private $externalTaxAmount;
 
     /**
+
+     * @var ?string
+     */
+    private $shippingKey;
+
+    /**
      * <p><code>id</code> of the <a href="ctp:api:type:CustomLineItem">CustomLineItem</a> to update. Either <code>customLineItemId</code> or <code>customLineItemKey</code> is required.</p>
      *
 
@@ -70,6 +76,18 @@ final class CartSetCustomLineItemTaxAmountActionBuilder implements Builder
     public function getExternalTaxAmount()
     {
         return $this->externalTaxAmount instanceof ExternalTaxAmountDraftBuilder ? $this->externalTaxAmount->build() : $this->externalTaxAmount;
+    }
+
+    /**
+     * <p><code>key</code> of the <a href="ctp:api:type:ShippingMethod">ShippingMethod</a> used for this Custom Line Item.
+     * This is required for Carts with <code>Multiple</code> <a href="ctp:api:type:ShippingMode">ShippingMode</a>.</p>
+     *
+
+     * @return null|string
+     */
+    public function getShippingKey()
+    {
+        return $this->shippingKey;
     }
 
     /**
@@ -106,6 +124,17 @@ final class CartSetCustomLineItemTaxAmountActionBuilder implements Builder
     }
 
     /**
+     * @param ?string $shippingKey
+     * @return $this
+     */
+    public function withShippingKey(?string $shippingKey)
+    {
+        $this->shippingKey = $shippingKey;
+
+        return $this;
+    }
+
+    /**
      * @deprecated use withExternalTaxAmount() instead
      * @return $this
      */
@@ -121,7 +150,8 @@ final class CartSetCustomLineItemTaxAmountActionBuilder implements Builder
         return new CartSetCustomLineItemTaxAmountActionModel(
             $this->customLineItemId,
             $this->customLineItemKey,
-            $this->externalTaxAmount instanceof ExternalTaxAmountDraftBuilder ? $this->externalTaxAmount->build() : $this->externalTaxAmount
+            $this->externalTaxAmount instanceof ExternalTaxAmountDraftBuilder ? $this->externalTaxAmount->build() : $this->externalTaxAmount,
+            $this->shippingKey
         );
     }