+++ lib/commercetools-api/src/Models/OrderEdit/StagedOrderSetCustomLineItemTaxAmountAction.php
@@ -18,6 +18,7 @@ interface StagedOrderSetCustomLineItemTaxAmountAction extends StagedOrderUpdateA
     public const FIELD_CUSTOM_LINE_ITEM_ID = 'customLineItemId';
     public const FIELD_CUSTOM_LINE_ITEM_KEY = 'customLineItemKey';
     public const FIELD_EXTERNAL_TAX_AMOUNT = 'externalTaxAmount';
+    public const FIELD_SHIPPING_KEY = 'shippingKey';
 
     /**
      * <p><code>id</code> of the <a href="ctp:api:type:CustomLineItem">CustomLineItem</a> to update. Either <code>customLineItemId</code> or <code>customLineItemKey</code> is required.</p>
@@ -45,6 +46,15 @@ interface StagedOrderSetCustomLineItemTaxAmountAction extends StagedOrderUpdateA
     public function getExternalTaxAmount();
 
     /**
+     * <p><code>key</code> of the <a href="ctp:api:type:ShippingMethod">ShippingMethod</a> used for this Custom Line Item.
+     * This is required for Carts with <code>Multiple</code> <a href="ctp:api:type:ShippingMode">ShippingMode</a>.</p>
+     *
+
+     * @return null|string
+     */
+    public function getShippingKey();
+
+    /**
      * @param ?string $customLineItemId
      */
     public function setCustomLineItemId(?string $customLineItemId): void;
@@ -58,4 +68,9 @@ interface StagedOrderSetCustomLineItemTaxAmountAction extends StagedOrderUpdateA
      * @param ?ExternalTaxAmountDraft $externalTaxAmount
      */
     public function setExternalTaxAmount(?ExternalTaxAmountDraft $externalTaxAmount): void;
+
+    /**
+     * @param ?string $shippingKey
+     */
+    public function setShippingKey(?string $shippingKey): void;
 }