+++ lib/commercetools-api/src/Models/OrderEdit/StagedOrderSetCustomLineItemTaxAmountActionModel.php
@@ -48,6 +48,12 @@ final class StagedOrderSetCustomLineItemTaxAmountActionModel extends JsonObjectM
      */
     protected $externalTaxAmount;
 
+    /**
+     *
+     * @var ?string
+     */
+    protected $shippingKey;
+
 
     /**
      * @psalm-suppress MissingParamType
@@ -56,11 +62,13 @@ final class StagedOrderSetCustomLineItemTaxAmountActionModel extends JsonObjectM
         ?string $customLineItemId = null,
         ?string $customLineItemKey = null,
         ?ExternalTaxAmountDraft $externalTaxAmount = null,
+        ?string $shippingKey = null,
         ?string $action = null
     ) {
         $this->customLineItemId = $customLineItemId;
         $this->customLineItemKey = $customLineItemKey;
         $this->externalTaxAmount = $externalTaxAmount;
+        $this->shippingKey = $shippingKey;
         $this->action = $action ?? self::DISCRIMINATOR_VALUE;
     }
 
@@ -144,6 +152,27 @@ final class StagedOrderSetCustomLineItemTaxAmountActionModel extends JsonObjectM
         return $this->externalTaxAmount;
     }
 
+    /**
+     * <p><code>key</code> of the <a href="ctp:api:type:ShippingMethod">ShippingMethod</a> used for this Custom Line Item.
+     * This is required for Carts with <code>Multiple</code> <a href="ctp:api:type:ShippingMode">ShippingMode</a>.</p>
+     *
+     *
+     * @return null|string
+     */
+    public function getShippingKey()
+    {
+        if (is_null($this->shippingKey)) {
+            /** @psalm-var ?string $data */
+            $data = $this->raw(self::FIELD_SHIPPING_KEY);
+            if (is_null($data)) {
+                return null;
+            }
+            $this->shippingKey = (string) $data;
+        }
+
+        return $this->shippingKey;
+    }
+
 
     /**
      * @param ?string $customLineItemId
@@ -167,5 +196,13 @@ final class StagedOrderSetCustomLineItemTaxAmountActionModel extends JsonObjectM
     public function setExternalTaxAmount(?ExternalTaxAmountDraft $externalTaxAmount): void
     {
         $this->externalTaxAmount = $externalTaxAmount;
+    }
+
+    /**
+     * @param ?string $shippingKey
+     */
+    public function setShippingKey(?string $shippingKey): void
+    {
+        $this->shippingKey = $shippingKey;
     }
 }